finish the video, tinggal finishing gess

+++ app/Http/Controllers/ListingController.php
@@ -50,6 +50,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing berhasil dibuat!!');
@@ -64,6 +66,10 @@ class ListingController extends Controller
     // update listing data
     public function update(Request $request, Listing $listing)
     {
+        // make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $formFields = $request->validate([
             "title" => "required",
             "company" => ["required"],
@@ -83,10 +89,21 @@ class ListingController extends Controller
         return back()->with('message', 'Listing berhasil diedit!!');
     }
 
+    // Delete Listing
     public function destroy(Listing $listing)
     {
+        // make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing berhasil dihapus');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
