+++ routes/web.php
@@ -3,8 +3,9 @@
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use App\Http\Controllers\BelajarWeb;
-use App\Http\Controllers\ListingController;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 
 /*
 |--------------------------------------------------------------------------
@@ -30,35 +31,37 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [ListingController::class, 'index']);
 
 // Create Form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store Listing Data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Show Edit Form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update Listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // Single Listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
-// Just Try again
+// Show User Registration/create form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
-// Route::get('/hello', function () {
-//     return view('hello');
-// });
+// Create a new user
+Route::post('/users', [UserController::class, 'store']);
 
-// Route::get('/posts/{id}', function ($id) {
-//     ddd($id);
-//     return response('Post of ' . $id);
-// })->where('id', '[0-9]+');
+// User Logout
+Route::post('/logout', [UserController::class, 'logout']);
 
-// Route::get('/search', function (Request $request) {
-//     dd($request->name . ' ' . $request->city . ' ' . $request->age);
-// });
+// Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// Login user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
